Wiki: improve readability for the 4.0 upgrade guides
The GitHub wiki markdown parser does not handle new lines in the same way as the markdown parser used, for instance, for GitHub issues.

The net effect of this, is that when there is a blank line between two sentences, the sentences will be formatted as separate paragraphs, but if there is only a new line between two sentences, they will be joined together into one long paragraph.
So using a new line will not enforce a "soft break" (`<br/>`) within a paragraph.

To get a "soft break" with the GH wiki markdown parser, you need two trailing spaces at the end of a sentence.
The trouble with that, is, of course, that a lot of editors will automatically try to remove these kind of trailing spaces, but so be it. Just means reviewers need to be extra careful about this.